Moved source file Vector.FancyIndexing.cs into the LinearAlgebra subfolder Implemented read-only logical indexing
